MedHack2020
In the midst of this pandemic COV19, Being in the queue sometimes in hospitals or clinic can be avoided by using this software which allow the patients to
book appointment ahead of time and come to meet the doctor at the appointed time to avoid crowd and being in the queue.The solutions this software is bringing to hospitals and clinic is to reduce contact between patients and healthcare practitioners in the midst of COV19.
The software can be used to manage patients appointments, know the number of patients to attend to per day,track unattended patients,
manage the space at the sitting room and many more.
This could save time for patients as well.